Skip merging when pages per sheet equals to 1

+++ main.py
@@ -15,18 +15,19 @@ def parsing():
     return parser.parse_args()
 
 
-def main(url, output_folder):
+def main(url, output_folder, pages_per_sheet=4):
     if 'file/d' in url:
         download_single_file(url, output_folder)
     if 'drive/folders' in url:
         download_a_folder(url, output_folder)
 
-    for file in glob(f'{output_folder}/*.pdf'):
-        if not '4_in_1' in file:
-            merge_pages(file, f'{file.split(".pdf")[0]}_4_in_1.pdf', 4)
-            if os.path.exists(file):
-                os.remove(file)
-            print(f'{file} has been merged')
+    if pages_per_sheet != 1:
+        for file in glob(f'{output_folder}/*.pdf'):
+            if not '4_in_1' in file:
+                merge_pages(file, f'{file.split(".pdf")[0]}_4_in_1.pdf', pages_per_sheet)
+                if os.path.exists(file):
+                    os.remove(file)
+                print(f'{file} has been merged')
 
 
 if __name__ == '__main__':
